+++ shared/src/index.ts
@@ -1,4 +1,4 @@
-export const GAME_VERSION = "0.205.1";
+export const GAME_VERSION = "0.205.2";
 
 export const TICK_RATE = 20;